requirements.txt
requests
python-dotenv
pandas
pyarrow
duckdb
jupyter
notebook